Refactoring

+++ ab/nn/util/ZST.py
@@ -1,4 +1,3 @@
-import os
 from pathlib import Path
 
 import zstandard as zstd
@@ -10,20 +9,12 @@ def compress(input_path: Path, output_path: Path, remove: bool = False):
     )
     with open(input_path, "rb") as fin, open(output_path, "wb") as f:
         compressor.copy_stream(fin, f)
-    if remove:
-        try:
-            input_path.unlink()
-        except Exception as e:
-            print(e)
+    if remove: input_path.unlink()
+
 
 def decompress(input_path: Path, output_path: Path, remove: bool = False):
     decompressor = zstd.ZstdDecompressor()
 
     with open(input_path, "rb") as fin, open(output_path, "wb") as f:
         decompressor.copy_stream(fin, f)
-    if remove:
-        if remove:
-            try:
-                input_path.unlink()
-            except Exception as e:
-                print(e)
+    if remove: input_path.unlink()
